+++ ReleaseNotes.md
@@ -49,3 +49,4 @@
 - Add MileGetFileAttributesByHandle function.
 - Add MileSetFileAttributesByHandle function.
 - Add MileGetFileHardlinkCountByHandle function.
+- Add MileDeleteFileByHandle function.